refactor: scatter mark utils
